+++ packages/flutter_state_notifier/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+- Add support for `LocatorMixin.initState` on `StateNotifierProvider`
+
 ## 0.2.0
 
 - Added `.value` contructor on `StateNotifierProvider`