+++ packages/state_notifier/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.3.0] - 08/03/2020
+
+Add an `initState` life-cycle to be able to init the `StateNotifier` using `read`.
+
 ## [0.1.0+1] - 06/03/2020
 
 Add example